--- a/docker/Makefile.am
+++ b/docker/Makefile.am
@@ -2,15 +2,17 @@ DCMP = docker-compose
 COMPOSE_SOURCE = docker-compose.yml elasticsearch/Dockerfile logstash/Dockerfile logstash/config/logstash.conf logstash/templates/wforce_template.json regression/Dockerfile
 COMPOSE_TARGET = .docker
 GEOIP_FILENAME = GeoLite2-City.mmdb
-GEOIP_FILE_GZ = logstash/geoip/$(GEOIP_FILENAME).gz
+GEOIP_FILE_GZ = logstash/geoip/$(GEOIP_FILENAME).tar.gz
 GEOIP_FILE = logstash/geoip/$(GEOIP_FILENAME)
 REGRESSION_SERVICE = regression
 
 $(GEOIP_FILE_GZ):
-	wget -N -O $(GEOIP_FILE_GZ) http://geolite.maxmind.com/download/geoip/database/$(GEOIP_FILENAME).gz
+	wget -N -O $(GEOIP_FILE_GZ) "https://download.maxmind.com/app/geoip_download?edition_id=GeoLite2-City&license_key=SPdi9BhAhMj6oOpw&suffix=tar.gz"
+
 
 $(GEOIP_FILE): $(GEOIP_FILE_GZ)
-	gunzip -c $(GEOIP_FILE_GZ) >$(GEOIP_FILE)
+	gunzip -c $(GEOIP_FILE_GZ) | tar xvf -
+	mv GeoLite2*/GeoLite2-City.mmdb $(GEOIP_FILE)
 
 $(COMPOSE_TARGET): $(COMPOSE_SOURCE) $(GEOIP_FILE) $(shell find ../common -type f) $(shell find ../wforce -type f) $(shell find ../trackalert -type f) $(shell find ../ext -type f) $(shell find ../regression-tests -type f)
 	$(DCMP) down -v