--- a/docker/logstash/Dockerfile
+++ b/docker/logstash/Dockerfile
@@ -4,9 +4,10 @@ USER root
 RUN yum install -y wget
 RUN mkdir -p /usr/share/logstash/geoip
 RUN chown -R logstash:logstash /usr/share/logstash/geoip
-RUN wget http://geolite.maxmind.com/download/geoip/database/GeoLite2-City.mmdb.gz
-RUN gunzip GeoLite2-City.mmdb.gz
-RUN mv GeoLite2-City.mmdb /usr/share/logstash/geoip
+RUN wget -O GeoLite2-City.mmdb.tar.gz "https://download.maxmind.com/app/geoip_download?edition_id=GeoLite2-City&license_key=SPdi9BhAhMj6oOpw&suffix=tar.gz"
+RUN gunzip GeoLite2-City.mmdb.tar.gz
+RUN tar xvf GeoLite2-City.mmdb.tar
+RUN mv GeoLite2*/GeoLite2-City.mmdb /usr/share/logstash/geoip
 USER logstash
 
 RUN logstash-plugin install logstash-input-udp