--- a/regression-tests/wforce-tw.conf
+++ b/regression-tests/wforce-tw.conf
@@ -12,12 +12,14 @@ field_map["countryCount"] = "countmin"
 newStringStatsDB("15SecondsFirstDB",1,15,field_map)
 sdb = getStringStatsDB("15SecondsFirstDB")
 sdb:twEnableReplication()
+sdb:twSetExpireSleep(250)
 
 newStringStatsDB("15SecondsSecondDB",1,15,field_map)
 sdb_prefix = getStringStatsDB("15SecondsSecondDB")
 sdb_prefix:twSetv4Prefix(24)
 sdb_prefix:twSetv6Prefix(64)
 sdb_prefix:twEnableReplication()
+sdb_prefix:twSetExpireSleep(250)
 
 newStringStatsDB("15SecondsSmallDB",1,15,field_map)
 sdb_small = getStringStatsDB("15SecondsSmallDB")