--- a/regression-tests/wforce-tw.conf
+++ b/regression-tests/wforce-tw.conf
@@ -9,19 +9,19 @@ field_map["diffPasswords"] = "hll"
 field_map["diffIPs"] = "hll"
 field_map["countryCount"] = "countmin"
 
-newStringStatsDB("15SecondsFirstDB",1,15,field_map)
+newShardedStringStatsDB("15SecondsFirstDB",1,15,field_map, 10)
 sdb = getStringStatsDB("15SecondsFirstDB")
 sdb:twEnableReplication()
 sdb:twSetExpireSleep(250)
 
-newStringStatsDB("15SecondsSecondDB",1,15,field_map)
+newShardedStringStatsDB("15SecondsSecondDB",1,15,field_map, 10)
 sdb_prefix = getStringStatsDB("15SecondsSecondDB")
 sdb_prefix:twSetv4Prefix(24)
 sdb_prefix:twSetv6Prefix(64)
 sdb_prefix:twEnableReplication()
 sdb_prefix:twSetExpireSleep(250)
 
-newStringStatsDB("15SecondsSmallDB",1,15,field_map)
+newShardedStringStatsDB("15SecondsSmallDB",1,15,field_map, 10)
 sdb_small = getStringStatsDB("15SecondsSmallDB")
 sdb_small:twSetMaxSize(10)
 sdb_small:twEnableReplication()