--- a/wforce/twmap-wrapper.cc
+++ b/wforce/twmap-wrapper.cc
@@ -25,17 +25,26 @@
 #include "replication_sdb.hh"
 #include "replication.pb.h"
 
-TWStringStatsDBWrapper::TWStringStatsDBWrapper(const std::string& name, int window_size, int num_windows)
+TWStringStatsDBWrapper::TWStringStatsDBWrapper(const std::string& name, int w_size, int n_windows, int n_shards) : window_size(w_size), num_windows(n_windows), num_shards(n_shards), db_name(name)
 {
-  sdbp = std::make_shared<TWStatsDB<std::string>>(name, window_size, num_windows);
+  sdbvp = std::make_shared<std::vector<std::shared_ptr<TWStatsDB<std::string>>>>();
+  for (int i=0; i<num_shards; ++i) {
+    sdbvp->push_back(std::make_shared<TWStatsDB<std::string>>(name, w_size, n_windows));
+  }
   replicated = std::make_shared<bool>(false);
 }
 
-TWStringStatsDBWrapper::TWStringStatsDBWrapper(const std::string& name, int window_size, int num_windows, const std::vector<pair<std::string, std::string>>& fmvec)
+TWStringStatsDBWrapper::TWStringStatsDBWrapper(const std::string& name, int window_size, int num_windows) : TWStringStatsDBWrapper(name, window_size, num_windows, 1)
 {
-  sdbp = std::make_shared<TWStatsDB<std::string>>(name, window_size, num_windows);    
-  replicated = std::make_shared<bool>(false);
-  (void)setFields(fmvec);
+}
+
+TWStringStatsDBWrapper::TWStringStatsDBWrapper(const std::string& name, int window_size, int num_windows, const std::vector<pair<std::string, std::string>>& fmvec, int num_shards) : TWStringStatsDBWrapper(name, window_size, num_windows, num_shards)
+{
+ (void)setFields(fmvec);
+}
+
+TWStringStatsDBWrapper::TWStringStatsDBWrapper(const std::string& name, int window_size, int num_windows, const std::vector<pair<std::string, std::string>>& fmvec) : TWStringStatsDBWrapper(name, window_size, num_windows, fmvec, 1)
+{
 }
 
 void TWStringStatsDBWrapper::enableReplication()
@@ -55,38 +64,49 @@ bool TWStringStatsDBWrapper::getReplicationStatus()
 
 std::string TWStringStatsDBWrapper::getDBName()
 {
-  return sdbp->getDBName();
+  return db_name;
 }
 
 bool TWStringStatsDBWrapper::setFields(const std::vector<pair<std::string, std::string>>& fmvec)
 {
-  FieldMap fm;
-  for(const auto& f : fmvec) {
-    fm.insert(std::make_pair(f.first, f.second));
+  for (const auto& sdbp : *sdbvp) {
+    FieldMap fm;
+    for(const auto& f : fmvec) {
+      fm.insert(std::make_pair(f.first, f.second));
+    }
+    sdbp->setFields(fm);
   }
-  return sdbp->setFields(fm);
+  return true;
 }
 
 bool TWStringStatsDBWrapper::setFields(const FieldMap& fm) 
 {
-  return(sdbp->setFields(fm));
+  for (const auto& sdbp : *sdbvp) {
+    sdbp->setFields(fm);
+  }
+  return true;
 }
 
 const FieldMap& TWStringStatsDBWrapper::getFields()
 {
-  return(sdbp->getFields());
+  // Just return the fields from the first
+  return(sdbvp->at(0)->getFields());
 }
 
 void TWStringStatsDBWrapper::setv4Prefix(uint8_t bits)
 {
   uint8_t v4_prefix = bits > 32 ? 32 : bits;
-  sdbp->setv4Prefix(v4_prefix);
+  for (const auto& sdbp : *sdbvp) {
+    sdbp->setv4Prefix(v4_prefix);
+  }
 }
 
 void TWStringStatsDBWrapper::setv6Prefix(uint8_t bits)
 {
   uint8_t v6_prefix = bits > 128 ? 128 : bits;
-  sdbp->setv6Prefix(v6_prefix);
+  for (const auto& sdbp : *sdbvp) {
+    sdbp->setv6Prefix(v6_prefix);
+  }
 }
 
 std::string TWStringStatsDBWrapper::getStringKey(const TWKeyType vkey)
@@ -100,11 +120,11 @@ std::string TWStringStatsDBWrapper::getStringKey(const TWKeyType vkey)
   else if (vkey.which() == 2) {
     const ComboAddress ca =  boost::get<ComboAddress>(vkey);
     if (ca.isIpv4()) {
-      uint8_t v4_prefix = sdbp->getv4Prefix();
+      uint8_t v4_prefix = sdbvp->at(0)->getv4Prefix();
       return Netmask(ca, v4_prefix).toStringNetwork();
     }
     else if (ca.isIpv6()) {
-      uint8_t v6_prefix = sdbp->getv6Prefix();
+      uint8_t v6_prefix = sdbvp->at(0)->getv6Prefix();
       return Netmask(ca, v6_prefix).toStringNetwork();
     }
   }
@@ -116,11 +136,20 @@ void TWStringStatsDBWrapper::add(const TWKeyType vkey, const std::string& field_
   addInternal(vkey, field_name, param1, param2, true);
 }
 
+const uint32_t tw_hash_seed = 623;
+
+unsigned int TWStringStatsDBWrapper::getShardIndex(const std::string& key)
+{
+  uint32_t hash;
+  MurmurHash3_x86_32(key.c_str(), key.length(), tw_hash_seed, (void*) &hash);
+  return hash % num_shards;
+}
+
 void TWStringStatsDBWrapper::addInternal(const TWKeyType vkey, const std::string& field_name, const boost::variant<std::string, int, ComboAddress>& param1, boost::optional<int> param2, bool replicate)
 {	
   std::string key = getStringKey(vkey);
   std::shared_ptr<SDBReplicationOperation> sdb_rop;
-  std::string db_name = sdbp->getDBName();
+  std::shared_ptr<TWStatsDB<std::string>> sdbp = sdbvp->at(getShardIndex(key));
   
   // we're using the three argument version
   if (param2) {
@@ -176,7 +205,7 @@ void TWStringStatsDBWrapper::subInternal(const TWKeyType vkey, const std::string
 {
   std::string key = getStringKey(vkey);
   std::shared_ptr<SDBReplicationOperation> sdb_rop;
-  std::string db_name = sdbp->getDBName();
+  std::shared_ptr<TWStatsDB<std::string>> sdbp = sdbvp->at(getShardIndex(key));
 
   if (val.which() == 0) {
     std::string mystr = boost::get<std::string>(val);
@@ -200,6 +229,8 @@ void TWStringStatsDBWrapper::subInternal(const TWKeyType vkey, const std::string
 int TWStringStatsDBWrapper::get(const TWKeyType vkey, const std::string& field_name, const boost::optional<boost::variant<std::string, ComboAddress>> param1)
 {
   std::string key = getStringKey(vkey);
+  std::shared_ptr<TWStatsDB<std::string>> sdbp = sdbvp->at(getShardIndex(key));
+
   if (param1) {
     if (param1->which() == 0) {
       std::string s = boost::get<std::string>(*param1);  
@@ -218,6 +249,8 @@ int TWStringStatsDBWrapper::get(const TWKeyType vkey, const std::string& field_n
 int TWStringStatsDBWrapper::get_current(const TWKeyType vkey, const std::string& field_name, const boost::optional<boost::variant<std::string, ComboAddress>> param1)
 {
   std::string key = getStringKey(vkey);
+  std::shared_ptr<TWStatsDB<std::string>> sdbp = sdbvp->at(getShardIndex(key));
+
   if (param1) {
     if (param1->which() == 0) {
       std::string s = boost::get<std::string>(*param1);  
@@ -238,6 +271,8 @@ std::vector<int> TWStringStatsDBWrapper::get_windows(const TWKeyType vkey, const
   std::vector<int> retvec;
     
   std::string key = getStringKey(vkey);
+  std::shared_ptr<TWStatsDB<std::string>> sdbp = sdbvp->at(getShardIndex(key));
+
   if (param1) {
     if (param1->which() == 0) {
       std::string s = boost::get<std::string>(*param1);  
@@ -257,6 +292,7 @@ std::vector<int> TWStringStatsDBWrapper::get_windows(const TWKeyType vkey, const
 bool TWStringStatsDBWrapper::get_all_fields(const TWKeyType vkey, std::vector<std::pair<std::string, int>>& ret_vec)
 {
   std::string key = getStringKey(vkey);
+  std::shared_ptr<TWStatsDB<std::string>> sdbp = sdbvp->at(getShardIndex(key));
   return sdbp->get_all_fields(key, ret_vec);
 }
 
@@ -269,7 +305,7 @@ void TWStringStatsDBWrapper::resetInternal(const TWKeyType vkey, bool replicate)
 {
   std::string key = getStringKey(vkey);
   std::shared_ptr<SDBReplicationOperation> sdb_rop;
-  std::string db_name = sdbp->getDBName();
+  std::shared_ptr<TWStatsDB<std::string>> sdbp = sdbvp->at(getShardIndex(key));
 
   sdbp->reset(key);
 
@@ -290,7 +326,7 @@ void TWStringStatsDBWrapper::resetFieldInternal(const TWKeyType vkey, const std:
 {
   std::string key = getStringKey(vkey);
   std::shared_ptr<SDBReplicationOperation> sdb_rop;
-  std::string db_name = sdbp->getDBName();
+  std::shared_ptr<TWStatsDB<std::string>> sdbp = sdbvp->at(getShardIndex(key));
 
   sdbp->reset_field(key, field_name);
 
@@ -305,63 +341,94 @@ void TWStringStatsDBWrapper::resetFieldInternal(const TWKeyType vkey, const std:
 
 unsigned int TWStringStatsDBWrapper::get_size()
 {	
-  return sdbp->get_size();
+  unsigned int size=0;
+  for (const auto& sdbp : *sdbvp) {
+    size += sdbp->get_size();
+  }
+  return size;
 }
 
 void TWStringStatsDBWrapper::set_size_soft(unsigned int size) 
 {
-  sdbp->set_map_size_soft(size);
+  for (const auto& sdbp : *sdbvp) {
+    sdbp->set_map_size_soft(size/num_shards);
+  }
 }
 
 void TWStringStatsDBWrapper::set_expire_sleep(unsigned int ms)
 {
-  sdbp->set_expire_sleep(ms);
+  for (const auto& sdbp : *sdbvp) {
+    sdbp->set_expire_sleep(ms);
+  }
 }
 
 unsigned int TWStringStatsDBWrapper::get_max_size()
-{	
-  return sdbp->get_max_size();
+{
+  unsigned int max_size=0;
+  for (const auto& sdbp : *sdbvp) {
+    max_size += sdbp->get_max_size();
+  }
+  return max_size;
 }
-
+    
 int TWStringStatsDBWrapper::windowSize()
 {
-  return sdbp->windowSize();
+  return window_size;
 }
 
 int TWStringStatsDBWrapper::numWindows()
 {
-  return sdbp->numWindows();
+  return num_windows;
 }
 
-const std::list<std::string>::iterator TWStringStatsDBWrapper::startDBDump()
+int TWStringStatsDBWrapper::numShards()
 {
-  return sdbp->startDBDump();
+  return num_shards;
 }
 
-bool TWStringStatsDBWrapper::DBDumpEntry(std::list<std::string>::iterator& i,
+VTWPtr::const_iterator TWStringStatsDBWrapper::begin() const
+{
+  return sdbvp->cbegin();
+}
+
+VTWPtr::const_iterator TWStringStatsDBWrapper::end() const
+{
+  return sdbvp->cend();
+}
+
+std::list<std::string>::const_iterator TWStringStatsDBWrapper::startDBDump(VTWPtr::const_iterator& sdbp) const
+{
+  return (*sdbp)->startDBDump();
+}
+
+bool TWStringStatsDBWrapper::DBDumpEntry(VTWPtr::const_iterator& sdbp,
+                                         std::list<std::string>::const_iterator& i,
                                          TWStatsDBDumpEntry& entry,
-                                         std::string& key)
+                                         std::string& key) const
 {
-  return sdbp->DBDumpEntry(i, entry, key);
+  return (*sdbp)->DBDumpEntry(i, entry, key);
 }
 
-const std::list<std::string>::iterator TWStringStatsDBWrapper::DBDumpIteratorEnd()
+const std::list<std::string>::const_iterator TWStringStatsDBWrapper::DBDumpIteratorEnd(VTWPtr::const_iterator& sdbp) const
 {
-  return sdbp->DBDumpIteratorEnd();
+  return (*sdbp)->DBDumpIteratorEnd();
 }
 
-void TWStringStatsDBWrapper::endDBDump()
+void TWStringStatsDBWrapper::endDBDump(VTWPtr::const_iterator& sdbp) const
 {
-  sdbp->endDBDump();
+  (*sdbp)->endDBDump();
 }
 
 void TWStringStatsDBWrapper::restoreEntry(const std::string& key, std::map<std::string, std::pair<std::time_t, TWStatsBufSerial>>& entry)
 {
+  std::shared_ptr<TWStatsDB<std::string>> sdbp = sdbvp->at(getShardIndex(key));
   sdbp->restoreEntry(key, entry);
 }
 
 void TWStringStatsDBWrapper::startExpireThread()
 {
-  thread t(TWStatsDB<std::string>::twExpireThread, sdbp);
-  t.detach();
+  for (auto& sdbp : *sdbvp) {
+    thread t(TWStatsDB<std::string>::twExpireThread, sdbp);
+    t.detach();
+  }
 }