… docs/manpages/wforce.1.md
… docs/manpages/wforce.conf.5.md