Three main improvements, all coming from experience at a very large customer with 30 million email subscribers:
1) Make the default statsdb expire/expunge thread sleep much shorter (30 secs -> 250ms), and also configurable. The 30s expiry sleep was causing noticeable spikes in latency causing sibling replication worker thread queues to fill up (this was the visible sign of the underlying issue). Doing expire more frequently means that most requests will take very slightly longer, but will avoid the big latency spikes every 30 seconds.
2) Make the sibling replication worker thread queue size configurable
3) Add an option to make stats dbs sharded. This creates multiple underlying stats dbs underneath the "hood" fo a single Lua statsdb. By doing so, it removes the single-lock contention and should make it much more scalable.

This PR includes regression tests and documentation for new config options